+++ go-app/internal/model/model.go
@@ -71,7 +71,7 @@ type Item struct {
 	CategoryID *string    `json:"category_id,omitempty" db:"category_id"` // カテゴリID（任意、外部キー）
 	UnitID     string     `json:"unit_id" db:"unit_id"`                   // 単位ID（必須、外部キー）
 	Quantity   *int       `json:"quantity,omitempty" db:"quantity"`       // 在庫数（任意）
-	UnitPrice  *float64   `json:"unit_price,omitempty" db:"unit_price"`   // 単価（円）
+	UnitPrice  *int       `json:"unit_price,omitempty" db:"unit_price"`   // 単価（円）
 	Status     string     `json:"status" db:"status"`                     // ステータス（active: 有効、inactive: 無効）
 	CreatedBy  *string    `json:"created_by,omitempty" db:"created_by"`   // 作成者のユーザーID（UUID、任意）
 	CreatedAt  time.Time  `json:"created_at" db:"created_at"`             // 作成日時
@@ -94,8 +94,8 @@ type StockHistory struct {
 	LocationTo   *string   `json:"location_to,omitempty" db:"location_to"`     // 移動先ロケーション（任意）
 	Reason       *string   `json:"reason,omitempty" db:"reason"`               // 理由・備考（任意）
 	Meta         string    `json:"meta" db:"meta"`                             // 補足情報（JSON形式）
-	UnitPrice    *float64  `json:"unit_price,omitempty" db:"unit_price"`       // 取引時の単価（円）
-	TotalAmount  *float64  `json:"total_amount,omitempty" db:"total_amount"`   // 取引金額（qty_delta × unit_price）
+	UnitPrice    *int      `json:"unit_price,omitempty" db:"unit_price"`       // 取引時の単価（円）
+	TotalAmount  *int      `json:"total_amount,omitempty" db:"total_amount"`   // 取引金額（qty_delta × unit_price）
 	CreatedBy    *string   `json:"created_by,omitempty" db:"created_by"`       // 実行者のユーザーID（任意）
 	CreatedAt    time.Time `json:"created_at" db:"created_at"`                 // 作成日時
 }
