+++ go-app/internal/repository/repository.go
@@ -761,7 +761,7 @@ func FetchStockHistory(limit, offset int) ([]model.StockHistory, int, error) {
 }
 
 // CreateItem はアイテムを作成します
-func CreateItem(code, name, unitID string, categoryID *string, quantity *int, unitPrice *float64) (*model.Item, error) {
+func CreateItem(code, name, unitID string, categoryID *string, quantity *int, unitPrice *int) (*model.Item, error) {
 	log.Printf("[Repository] CreateItem - code: %s, name: %s", code, name)
 
 	var item model.Item
@@ -792,7 +792,7 @@ func CreateItem(code, name, unitID string, categoryID *string, quantity *int, un
 }
 
 // UpdateItem はアイテムを更新します
-func UpdateItem(id, code, name, unitID string, categoryID *string, quantity *int, unitPrice *float64, status string) (*model.Item, error) {
+func UpdateItem(id, code, name, unitID string, categoryID *string, quantity *int, unitPrice *int, status string) (*model.Item, error) {
 	log.Printf("[Repository] UpdateItem - id: %s", id)
 
 	var item model.Item
